Move assignments types
